feat: consume user data
