+++ src/app/components/users/user-list/user-list.component.scss
@@ -1,17 +1,5 @@
-.user {
-  background: linear-gradient(186.5deg, #3c3e64 0.83%, #1d1e2a 100%);
-  box-shadow: 0 1px 0 #ffffff, 0px -1px 0px #ffffff;
-  color: #fff;
-  padding: 20px 40px;
-
-  display: flex;
-  align-items: center;
-}
-
-.user > .info-wrapper {
-  margin-left: 16px;
-}
-
-.user > .pay-button {
-  margin-left: auto;
+ul {
+  margin: 0;
+  padding: 0;
+  list-style: none;
 }
