--- src/app/components/users/user/user.component.html
@@ -1,21 +1,0 @@
-<div class="user" >
-  <img [src]="photo" alt="Avatar do usuário">
-  <div class="info-wrapper">
-    <div class="user-name">
-      <h1>
-        <ng-content></ng-content>
-        {{ photo.login }}
-      </h1>
-      
-      <p>{{ photo.login }}</p>
-    </div>
-
-    <div class="user-id-username">
-      ID: X - Username: @usuarioX
-    </div>
-  </div>
-
-  <button type="button" class="pay-button">
-    Pagar
-  </button>
-</div>
