--- src/app/components/users/user/user.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-user',
-  templateUrl: './user.component.html',
-  styleUrls: ['./user.component.scss']
-})
-export class UserComponent implements OnInit {
-  @Input() photo = '';
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
