+++ src/app/components/users/users.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 
-import { UserComponent } from './user/user.component';
 import { UserListComponent } from './user-list/user-list.component';
+import { UserComponent } from './user-list/user/user.component';
 
 @NgModule({
   declarations: [
